pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "pita"
version = "1.0.0"
description = "A library to allow probabilistic inference time algorithms from any model"
license = {text = "AGPL-3.0-or-later"}
readme = "README.md"
requires-python = ">=3.8"
authors = [
    {name = "COBI, Inc.", email = "sales@cobi-inc.com"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: GNU Affero General Public License v3 or later (AGPLv3+)",
    "License :: Other/Proprietary License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

# Core dependencies - minimal set needed for the library to work
dependencies = [
    "numpy>=1.20.0",
    "scipy>=1.10.0",
    "transformers>=4.30.0",
    "fastapi>=0.100.0",
    "uvicorn>=0.23.0",
    "pydantic>=2.0.0",
    "datasets",
    "pandas>=2.0.0",
    "psutil",
    "pylatexenc",
    "sympy",
    "huggingface-hub>=0.16.0",
    "click>=8.0.0",
]

[project.urls]
Homepage = "https://github.com/cobi-inc-MC/pita"
Repository = "https://github.com/cobi-inc-MC/pita"
Documentation = "https://github.com/cobi-inc-MC/pita/tree/main/docs"
"Bug Tracker" = "https://github.com/cobi-inc-MC/pita/issues"
"Commercial Licensing" = "https://github.com/cobi-inc-MC/pita/blob/main/docs/LICENSING-GUIDE.md"

[project.optional-dependencies]
# vLLM backend dependencies
vllm = [
    "vllm>=0.10.0",
    "valkey>=6.0.0",
]

# llama.cpp backend dependencies
llama_cpp = [
    "llama-cpp-python",
]

# TensorRT backend dependencies
tensorrt = [
    "tensorrt_llm",
    "valkey>=6.0.0",
]

# Development/testing dependencies
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
]

# Install all backends
all = [
    "pita[vllm]",
    "pita[llama_cpp]",
    "pita[tensorrt]",
    "pita[dev]",
]

[project.scripts]
pita = "pita.cli:cli"

[tool.setuptools]
packages = ["pita"]

[tool.pytest.ini_options]
norecursedirs = ["pita/old_tests"]